Refactored annotation method into more generic/reusable class
